# operations/transaction-service/openapi.yaml
openapi: 3.0.3
info:
  title: Transaction Service
  version: 1.0.0
  description: Client-scoped token-operations API for master wallet transfers and transaction queries.
servers:
  - url: /
security:
  - jwtAssertion: []
paths:
  /health:
    get:
      summary: Service liveness
      security: []
      responses:
        "200":
          description: Service is up
  /wallets:
    get:
      summary: List all wallets
      responses:
        "200":
          description: Wallet summaries, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WalletSummary"
        "401":
          $ref: "#/components/responses/Error"
  /wallets/addresses:
    get:
      summary: List all wallet addresses
      responses:
        "200":
          description: Wallet addresses, ordered by address
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "401":
          $ref: "#/components/responses/Error"
  /wallets/me:
    get:
      summary: List the caller's own wallets with balances
      description: >-
        Returns the wallets owned by the end user identified by the verified end-user
        token (userAssertion), with decrypted balances and the default wallet first.
        Requires both the service token (jwtAssertion) and a verified end-user token
        (userAssertion). Used by the People App to let a user pick which wallet to pay
        from.
      security:
        - jwtAssertion: []
          userAssertion: []
      responses:
        "200":
          description: The caller's wallets, default wallet first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserWallet"
        "401":
          $ref: "#/components/responses/Error"
  /wallets/master/balance:
    get:
      summary: Get the caller's master wallet balance
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
  /wallets/master/transfer:
    post:
      summary: Transfer from the caller's master wallet
      description: |
        Moves coins from the caller's master wallet to an existing wallet. A caller-supplied
        `reference` acts as the idempotency key: retrying with the same reference and identical
        details replays the recorded transaction (200) instead of moving coins again, while the
        same reference with different details is rejected (409). Without a reference the service
        generates one and every call is a new transfer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransferRequest"
      responses:
        "201":
          description: Transfer accepted
          headers:
            Location:
              description: URL of the created transaction
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "200":
          description: Idempotent replay of a transfer already recorded under the same reference
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
  /wallets/{address}/balance:
    get:
      summary: Get any wallet's balance
      parameters:
        - $ref: "#/components/parameters/Address"
      responses:
        "200":
          description: Balance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Balance"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /payments:
    post:
      summary: Collect a payment from a user's own wallet into a treasury wallet
      description: >-
        A scoped client collects a coin payment from the end user's own wallet into a
        treasury wallet on the user's behalf. Requires both the service token
        (jwtAssertion) and a verified end-user token (userAssertion). The reference is
        a caller-supplied idempotency key: replaying it with identical details returns
        the original transaction (200); reusing it with different details is a 409.
      security:
        - jwtAssertion: []
          userAssertion: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentRequest"
      responses:
        "200":
          description: Idempotent replay of an already-recorded payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "201":
          description: Payment recorded
          headers:
            Location:
              description: URL of the created transaction
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "403":
          $ref: "#/components/responses/Error"
        "409":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
  /transactions/search:
    post:
      summary: Browse and filter transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionSearchRequest"
      responses:
        "200":
          description: Transaction page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionPage"
        "400":
          $ref: "#/components/responses/Error"
        "401":
          $ref: "#/components/responses/Error"
        "413":
          $ref: "#/components/responses/Error"
  /transactions/{reference}:
    get:
      summary: Get a single transaction
      parameters:
        - $ref: "#/components/parameters/Reference"
      responses:
        "200":
          description: Transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "401":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    jwtAssertion:
      type: apiKey
      in: header
      name: X-Jwt-Assertion
      description: Gateway-issued JWT assertion carrying the caller identity.
    userAssertion:
      type: apiKey
      in: header
      name: X-User-Assertion
      description: >-
        Verified end-user JWT (header name configurable via USER_ASSERTION_HEADER).
        Its email claim identifies the wallet owner on whose behalf a payment is made.
  parameters:
    Address:
      name: address
      in: path
      required: true
      schema:
        type: string
    Reference:
      name: reference
      in: path
      required: true
      schema:
        type: string
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Balance:
      type: object
      properties:
        walletAddress:
          type: string
        balance:
          type: string
          description: Decimal balance as a string
    WalletSummary:
      type: object
      properties:
        walletAddress:
          type: string
        defaultWallet:
          type: boolean
        createdOn:
          type: string
          description: Creation timestamp in RFC3339 format
    UserWallet:
      type: object
      required: [walletAddress, balance, defaultWallet]
      properties:
        walletAddress:
          type: string
        balance:
          type: string
          description: Decimal balance as a string
        defaultWallet:
          type: boolean
    TransferRequest:
      type: object
      required: [toAddress, amount]
      properties:
        toAddress:
          type: string
        amount:
          type: string
          description: Decimal amount as a string
        reference:
          type: string
          maxLength: 66
          description: >-
            Optional caller-supplied idempotency key; matches ^[A-Za-z0-9][A-Za-z0-9._:-]{0,65}$.
            When omitted the service generates a 0x + 64 hex reference.
        source:
          type: string
          description: Optional uppercase source token; matches ^[A-Z][A-Z0-9_]{1,31}$
    TransferResponse:
      type: object
      properties:
        reference:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        amount:
          type: string
          description: Decimal amount as a string
    PaymentRequest:
      type: object
      required: [fromAddress, toAddress, amount, reference, source]
      properties:
        fromAddress:
          type: string
          description: The user-owned wallet the payment is collected from
        toAddress:
          type: string
          description: The treasury wallet the payment is collected into
        amount:
          type: string
          description: Decimal amount as a string; must be greater than zero
        reference:
          type: string
          maxLength: 66
          description: >-
            Caller-supplied idempotency key. Matches ^[A-Za-z0-9][A-Za-z0-9._:-]{0,65}$;
            the 0x + 64 hex shape used by service-generated references is also accepted.
        source:
          type: string
          description: Uppercase source token; matches ^[A-Z][A-Z0-9_]{1,31}$
    PaymentResponse:
      type: object
      properties:
        reference:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        amount:
          type: string
          description: Decimal amount as a string
    TransactionSearchRequest:
      type: object
      properties:
        fromAddresses:
          type: array
          items:
            type: string
        toAddresses:
          type: array
          items:
            type: string
        reference:
          type: string
        dateFrom:
          type: string
        dateTo:
          type: string
        limit:
          type: integer
          default: 15
          maximum: 100
        offset:
          type: integer
          default: 0
    Transaction:
      type: object
      properties:
        reference:
          type: string
        fromAddress:
          type: string
        toAddress:
          type: string
        amount:
          type: string
          description: Decimal amount as a string
        timestamp:
          type: string
    TransactionPage:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        limit:
          type: integer
        offset:
          type: integer
        next:
          type: string
          nullable: true
        previous:
          type: string
          nullable: true
    Error:
      type: object
      properties:
        message:
          type: string